# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "abcy-data API",
    "version": "1.0.0"
  },
  "paths": {
    "/activities": {
      "get": {
        "summary": "List recent activities",
        "parameters": [
          {
            "name": "count",
            "in": "query",
            "required": false,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {"description": "Activity list"}
        }
      }
    },
    "/activity/{id}": {
      "get": {
        "summary": "Full activity metadata and streams",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {"description": "Activity data"},
          "404": {"description": "Not found"}
        }
      }
    },
    "/activity/{id}/summary": {
      "get": {
        "summary": "Small summary for an activity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "responses": {
          "200": {
            "description": "Summary data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": {"type": "integer"},
                    "name": {"type": "string"},
                    "start_date": {"type": "string"},
                    "distance": {"type": "number"},
                    "duration": {"type": "integer"},
                    "weighted_average_power": {"type": "number"},
                    "average_speed": {"type": "number"},
                    "max_speed": {"type": "number"},
                    "pr_count": {"type": "integer"},
                    "average_heartrate": {"type": "number"},
                    "summary_polyline": {"type": "string"},
                    "normalized_power": {"type": "number"},
                    "intensity_factor": {"type": "number"},
                    "training_stress_score": {"type": "number"},
                    "activity_type": {"type": "string"},
                    "trend": {
                      "type": "object",
                      "properties": {
                        "avg_speed": {"type": "string"},
                        "max_speed": {"type": "string"},
                        "tss": {"type": "string"},
                        "intensity": {"type": "string"},
                        "power": {"type": "string"}
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {"description": "Not found"}
        }
      }
    },
    "/ftp": {
      "get": {"summary": "Current FTP", "responses": {"200": {"description": "Current FTP"}}},
      "post": {
        "summary": "Append new FTP value",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"type": "object", "properties": {"ftp": {"type": "number"}}, "required": ["ftp"]}
            }
          }
        },
        "responses": {"200": {"description": "Updated"}}
      }
    },
    "/weight": {
      "get": {"summary": "Current weight", "responses": {"200": {"description": "Current weight"}}},
      "post": {
        "summary": "Append new weight value",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {"type": "object", "properties": {"weight": {"type": "number"}}, "required": ["weight"]}
            }
          }
        },
        "responses": {"200": {"description": "Updated"}}
      }
    },
    "/wkg": {"get": {"summary": "Current W/kg", "responses": {"200": {"description": "Current W/kg"}}}},
    "/wkg/history": {"get": {"summary": "W/kg history", "responses": {"200": {"description": "History"}}}},
    "/enduro": {"get": {"summary": "Current EnduroScore", "responses": {"200": {"description": "Enduro"}}}},
    "/enduro/history": {"get": {"summary": "EnduroScore history", "responses": {"200": {"description": "History"}}}},
    "/fitness": {"get": {"summary": "Current FitnessScore", "responses": {"200": {"description": "Fitness"}}}},
    "/fitness/history": {"get": {"summary": "FitnessScore history", "responses": {"200": {"description": "History"}}}},
    "/trend": {"get": {"summary": "Recent trends", "responses": {"200": {"description": "Trends"}}}},
    "/stats": {
      "get": {
        "summary": "Aggregated statistics",
        "parameters": [
          {"name": "period", "in": "query", "required": true, "schema": {"type": "string", "enum": ["day","week","month","year"]}},
          {"name": "ids", "in": "query", "required": false, "schema": {"type": "string"}},
          {"name": "types", "in": "query", "required": false, "schema": {"type": "string"}}
        ],
        "responses": {"200": {"description": "Stats"}}
      }
    }
  }
}